+++ tests/test_mc_point.cpp
@@ -12,10 +12,11 @@ TYPED_TEST(McPoint, ConstructorAndMemberVariables)
 {
 	typedef TypeParam T;
 
-	std::vector<T> point = { T(0.1), T(0.9), T(0.133), T(0.4) };
-	hep::mc_point<T> result(1000, point);
+	std::vector<T> point{ T(0.1), T(0.9), T(0.133), T(0.4) };
+	hep::mc_point<T> result(point);
 
-	EXPECT_NEAR( T(0.001) , result.weight()   , T(1e-10) );
+	EXPECT_EQ( T(1.0) , result.weight() );
+
 	EXPECT_NEAR( T(0.1)   , result.point()[0] , T(1e-10) );
 	EXPECT_NEAR( T(0.9)   , result.point()[1] , T(1e-10) );
 	EXPECT_NEAR( T(0.133) , result.point()[2] , T(1e-10) );